+++ src/Sentry/Extensibility/IExceptionFilter.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Sentry.Internal
+namespace Sentry.Extensibility
 {
     /// <summary>
     /// A filter to be applied to an exception instance.
@@ -13,11 +13,5 @@ namespace Sentry.Internal
         /// <param name="ex">The exception about to be captured.</param>
         /// <returns><c>true</c> if [the event should be filtered out]; otherwise, <c>false</c></returns>.
         bool Filter(Exception ex);
-    }
-
-    internal class ExceptionTypeFilter<TException> : IExceptionFilter where TException : Exception
-    {
-        private readonly Type _filteredType = typeof(TException);
-        public bool Filter(Exception ex) => _filteredType.IsInstanceOfType(ex);
     }
 }
